requirements.txt
###### Requirements without Version Specifiers ######
pypiwin32
pycryptodome
